stowsh_add_to_package()
A new CLI option -a PACKAGE will take SOURCE command-line arguments,
and add these to the destination specified by -a PACKAGE. PACKAGE is
a path. The recommended way to use this towards a git repo is with
the -g option. If the repo is clean the sources may overwrite the
working tree and will also stage the new additions for review by the
user.