Add stowsh_list_status() command -l
This commit adds a function stowsh_list_status that lists the files in
SRCs and displays the corresponding file at the target if given by the
-t option or '.', otherwise.

If the user supplies the -g option a column is added to display the SRC
git status with characters as produced by the `git status --porcelain`
command.

This commit introduces a new dependency: `column` from the linux-util
package. It could be removed, by simplifying the report format when it
is not present.

Not thoroughly tested.