Option `-c` to continue adding to commit
When using the add `-a` command the `-c` option
will allows the user to continue to add files to the current commit.

The test of whether or not the path given with `-a` points into a repo is
moved into a new function git_is_path_inside_repo().

Testing of whether a repo is clean is now done only once per call.